added last updated vital next to graphs
