+++ src/components/patientAlertTab/AlertTabComponent.tsx
@@ -35,7 +35,7 @@ const AlertTabComponent = ({ patient }: PropType) => {
         ))}
       </div>
       <div>
-        {selectedTab === 0 && <PatientAlerts />}
+        {selectedTab === 0 && <PatientAlerts patient={patient} />}
         {selectedTab === 1 && <PatientConfigs patient={patient} />}
       </div>
     </div>
